+++ CHANGES.md
@@ -1,5 +1,11 @@
 ## CHANGELOG ##
 
+* v1.4.0 (2020-09-23)
+  * `name()` argument is now `trim()`ed automatically.
+  * Removed `invobill()` method. Use `reserved2()` instead.
+  * Corrected some tests.
+  * 100% test coverage.
+
 * v1.3.0 (2020-09-23)
   * Bumped min PHP version to 5.6
   * Added Travis-CI support.